Move the rwbuf code under util/.
It certainly doesn't need its own top-level package. What was I thinking?
