+++ src/python/pants/base/BUILD
@@ -172,9 +172,9 @@ python_library(
   sources = ['workunit.py'],
   dependencies = [
     '3rdparty/python:six',
-    'src/python/pants/rwbuf',
     'src/python/pants/util:dirutil',
     'src/python/pants/util:memo',
+    'src/python/pants/util:rwbuf',
   ],
 )
