--- src/python/pants/rwbuf/BUILD
@@ -1,10 +1,0 @@
-# Copyright 2014 Pants project contributors (see CONTRIBUTORS.md).
-# Licensed under the Apache License, Version 2.0 (see LICENSE).
-
-python_library(
-  name = 'rwbuf',
-  sources = globs('*.py'),
-  dependencies = [
-    '3rdparty/python:six',
-  ]
-)
