+++ src/python/pants/util/BUILD
@@ -56,6 +56,14 @@ python_library(
 )
 
 python_library(
+  name = 'rwbuf',
+  sources = ['rwbuf.py'],
+  dependencies = [
+    '3rdparty/python:six',
+  ]
+)
+
+python_library(
   name = 'strutil',
   sources = ['strutil.py'],
   dependencies = [
